requirements.txt
click==8.0.4
Flask==2.0.3
Flask-WTF==1.0.0
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysql-connector-python==8.0.28
protobuf==3.19.4
Werkzeug==2.0.3
WTForms==3.0.1
